backend/requirements.txt
# ═══════════════════════════════════════════════════════════════
# Mary Poppins — Backend Dependencies
# ═══════════════════════════════════════════════════════════════

# ── Core Framework ──
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.7,<3.0
pydantic-settings>=2.3,<3.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0

# ── Async HTTP / Networking ──
aiohttp>=3.10,<4.0
httpx>=0.27,<1.0
aiodns>=3.2,<4.0
aiofiles>=24.1,<25.0

# ── Databases ──
asyncpg>=0.29,<1.0
sqlalchemy[asyncio]>=2.0,<3.0
alembic>=1.13,<2.0
neo4j>=5.22,<6.0
elasticsearch[async]>=8.14,<9.0
redis[hiredis]>=5.0,<6.0

# ── Message Queue ──
celery[redis]>=5.4,<6.0
aiokafka>=0.11,<1.0
kombu>=5.3,<6.0

# ── AI / ML ──
onnxruntime>=1.18,<2.0
numpy>=1.26,<3.0
Pillow>=10.4,<12.0
imagehash>=4.3,<5.0

# ── OSINT ──
phonenumbers>=8.13,<9.0
asyncwhois>=1.1,<2.0

# ── Crypto Tracing ──
web3>=6.20,<7.0
bitcoin-utils>=0.6,<1.0

# ── Dark Web ──
aiohttp-socks>=0.9,<1.0
beautifulsoup4>=4.12,<5.0
lxml>=5.2,<6.0

# ── Geolocation ──
geoip2>=4.8,<5.0

# ── Security / Auth ──
python-keycloak>=4.2,<5.0
cryptography>=42.0,<44.0

# ── Monitoring / Observability ──
prometheus-client>=0.20,<1.0
prometheus-fastapi-instrumentator>=7.0,<8.0
structlog>=24.2,<25.0
sentry-sdk[fastapi]>=2.7,<3.0

# ── Utilities ──
python-dotenv>=1.0,<2.0
orjson>=3.10,<4.0
tenacity>=8.4,<9.0
click>=8.1,<9.0
